scripts/infra/gcp-billing-kill-switch/terraform/function_source/requirements.txt
google-cloud-billing==1.14.0
functions-framework==3.8.1
